DOCS AI/requirements.txt
langchain
PyPDF2
python-dotenv
streamlit
openai
faiss-cpu
altair
tiktoken
boto3
# unstructured
# unstructured[pdf]
pypdf
cryptography